Extract IndexedDataset into its own mod
